🐛Fix a major bug and implement review multiple choice.
The bug was comparing the user's answer against itself 🤦‍♀️. This falsely updated the wrong card in the database, that being the card to the button being clicked. It also updated this wrong card to have a correct answer. And it also didn't update the actual question's card at all 🤡

The fix was to use the question's ID as a value in the hidden field of the answer buttons, not the actual answer's ID.