Fix a bug with the calculation of next review date.
The bug: the next review date of a review card was set after setting the new ease. Additionally, the ease and the growth factor affected the next review date, but only ease is supposed to affect it.
The fix was to only use the ease to calculate next review date, and make sure it happens before setting the new ease after a card has been answered correctly.

Use dark theme.
Remove unused tests.